Work on the full body instead of the snippet - fixed.
Gmail has weird way of giving the response back via the API. Despite
them parsing the emails, they still want us to do the hard work of
checking here and there to get the emails.

So instead I took the raw email and used the stdlib to do the job with
beautifulsoup to parse out the text from html emails.

Does the job of getting us the body for now.