Streamline background job processing
Previously, the job would happen without using RunBatch. This meant that it
will be harder to parallelize in the future. Even though the processing right
now is going to be single job for each schedule, I am still making it such that
each batch of X messages will be a "batch" under that schedule. That way in the
future, when we want to branch this out, we have to write another job and just
hand it that batch and the rest of the processing remains the same