Accept # of emails to process
Because of OpenAI limits and the general sense of how long it takes to
process these emails, it will be better to just ask the user how many
emails they wish to process. The system then uses this to calculate how
many batches to receive and how many per batch.

It makes the assumption of rounding off to the nearest 500. So if the
user give 2,134 emails to process out of 3,000 then it will get 5
batches of 500. So it will actually process 2,500 emails. This is so
that we avoid the bookkeeping to ensure we process only 2,134 emails.