Set the job timeout to 1h
The documentation says this: job_timeout specifies the maximum runtime
of the job before it’s interrupted and marked as failed. This means that
irrespective of the job running, it will stop the job and kill it.
Weird.

So I am setting this to 1h for now. We need to decide what the optimum
time to set this should be.