Revert change made previously
It is very complex to handle job status related stuff on the home page
since there can be multiple jobs enqueued. So I am removing this from
the home page and deciding that I will handle it from the run status
page.

It is simpler in the run status page simply because there is only ONE
run available in context there.