Handle jobs properly (#4)
Abandoning the run will set the ended_at and then the status to
DONE_WITH_ERRORS. This will allow the user to schedule another run

It is very complex to handle job status related stuff on the home page
since there can be multiple jobs enqueued. So I am removing this from
the home page and deciding that I will handle it from the run status
page.

It is simpler in the run status page simply because there is only ONE
run available in context there.

# Handle refresh in pure JS

I was handling the refresh via HTML meta tags but because I need to be
able to pause and refresh the timer for refresh (since I am showing a
modal on the page), I decided to handle the refresh in JS entirely.